// client/src/app/core/models/game/card-type.ts
export enum CardType {
  UsedCard = 'UsedCard',
  FromCommonDeck = 'FromCommonDeck',
  Plus = 'Plus',
  Minus = 'Minus',
  PlusMinus = 'PlusMinus',
  OneOrTwoPlusMinus = 'OneOrTwoPlusMinus',
  ThreeOrFourPlusMinus = 'ThreeOrFourPlusMinus',
  FiveOrSixPlusMinus = 'FiveOrSixPlusMinus',
  AnyValue = 'AnyValue',
}